Support IFMap Dependency manager in AgentDB
IFMap Dependency manager changes expect that the IFMapNode is available
when an object is created. Added new class AgentOperDB which implements
OperDB entries supporting IFMap dependency tracking. Moved the
dependency tracking code from interface/vm-interface, physical-device
and VN to use the new infra

Call to IFNodeToReq copies IFMapNode pointer to AgentOperDBData structure. The
data is in-turn used passed in async request "Enqueue". Add a DBState to
IFMapNode to ensure that the IFMapNode is not deleted in the meantime

No new test cases added since its only a code re-org and there is no
functionality change.

Change-Id: Id14ebeaaa0ddc6f66994d75f5094bd65586b5c25