Prevent loops when re-originating service chain routes
Implement a new attribute called OriginVnPath to keep track of VNs from
which a given route has been re-originated. The OriginVn of ServiceChain
destination routing instance gets prepended to OriginVnPath when adding
the ServiceChain route. Further, a route is not re-originated if source
routing instance's OriginVn is already in the OriginVnPath.

This prevents route re-origination loops even if VNs are mis-configured
such that multiple VNs have the same route target and there are transit
VNs with mis-configured route targets as well.

Change-Id: Iab76ad4b07a69f71383db8e72531fb31bade04bd
Closes-Bug: 1401010