Adding of seperate EVPN table in agent along with L2(bridge) table. EVPN table is keyed with VRF, MAC, IP, ethernet_tag(vxlan/vnid). L2 table will be keyed only with MAC and VRF.
When a mac route is to be added either via EVPN or control node the request is
given to EVPN table. Same MAC can be attached to multiple IP e.g. V4 and v6 or
aliases. Each of them will be a sperate route in EVPN table and will be exported
to control node. This table is not linked to data plane programming i.e. ksync
does not listen to it. Each entry created in EVPN table creates a request to add
MAC in L2 table with a peer unique to the EVPN route. This peer is known as
EVPN_PEER and path added in L2 mac route is EVPN path. EVPN path is always the
active path of EVPN route.  Hence there will be one active path from each EVPN
route sharing same MAC. In turn it results in L2 MAC route having path for each
IP alias(or parent EVPN route).
EVPN paths will reference to agent path of parent(EVPN route).

L2 table is the one which will interface to ksync i.e. data plane. However this
will not be used for exporting route to control node. Exception is for multicast
route which will be shifted to EVPN route later. Its a TBD item.

Since now we have a EVPN table keyed with vxlan id, the logic of withdraw and
subscribe routes to control-node because of encap change or vxlan id change, is
no more required in controller_export code. It is handled via encap change and
vxlanid and mode change.

Assumption: For same MAC all IP addresses associated to it will have same NH.

Change-Id: I7a3a870792a0dd9379663b320ff20808ca9a1f90