Implement local AS knob for bgp-router
This allows a bgp-router to have a different AS than the global AS
configured for the VNS instance. It's useful in scenarios where the
VNS instance spans multiple geographical regions.

The local AS is used for peering, to prepend to the AS path, as the
AS in route target routes and for all other bgp procedures. The AS
from the VNS instance is used to generate route targets, security
groups etc.

Change-Id: Ie1a8bdd03b2e283f10a986cf4c1fd78dc2ffada2
Partial-Bug: 1408429