Allowed address configuration while creating the port.
While creating the port, the mac address is not yet allocated to port.
If the port's Allowed address pair config doesn't contain the mac_address, plugin tries to copy the port's mac for allowed address pair mac.
This results in exception

Fix:
1. Remove the overlap check.
   Code to check whether allowed address pair ip+mac doesn't match port's fixed-ip+mac.
   This check is no longer needed in neutron code. Ref:https://review.openstack.org/#/c/94508/

2. If the mac address is not specified in the allowed address pair config, then set the mac_address in allowed address pair config to ""(empty string)
Populate the mac address in API server for both PUT and POST

Change-Id: Ib59ae095a3fc12ae172d44511fc2d5b6e3df2edf
Closes-Bug: 1410631