Service monitor per message exception handling
Currently the top level exception handling in service monitor
exists for process_poll_result. The issue with this scheme is
that many IFMAP messages are clumped together and hence any
exception on an intermediate message stops the processing of
the rest of the messages in the queue.  Exception should be
caught on a per message basis.

Change-Id: I3190fc71b3f87fbaf2d11a7229de881b5957c3c0
Closes-Bug: #1413787