Force RESYNC on VMI added from config
For VM Interface, the VN and VRF link are built when RESYNC is called.
If the link from sub-interface VMI to parent interface VMI is received
last, only Add-Change is called for the VMI sub-interface.

Modifying PostAdd() for VMInterface to force Change (RESYNC in-turn) for
VMInterface

Add UT for the same.

Change-Id: I6e074501fb10d7d1b48f86111fd0ef086c69aa56