config-openstack: do not maintain internal cache for proj/domain
An internal cache was maintained earlier for projects and domains
to optimize on keystone poll for adds. Since we have effectively
moved to add project/domain on demand, there is nothing to be had
by maintaining internal cache (with the invalidation/resyncing
problems). Instead read the project/domain from contrail database
on pre_<proj|domain>_read and sync only if not already present.

Change-Id: I43193ac3c15c8cd3e07fd59bfd5ea3b7770c78b3