Add more specific rules for VRF assign table.
For service instance interfaces, we use VRF assign table to assign transit
traffic to service RI. Traffic originating in the service instance is supposed
to be assigned to native RI. This was done by adding another rule based on SIP.
However, if floating ip is assigned to service instance, the traffic with FIP
will not match the SIP rule and will match the rule for transit traffic. Now we
add source and destination VN to the rule for transit traffic so that other
traffic will not match this rule.

Change-Id: I21d0fcc3abd66e7a7c64d015db4142bda9e64f08
Closes-Bug: 1394259