Nexthop notification server functionality in vrouter agent
Enable vrouter agent to send notification about tunnel nexthops to interested clients. This is implemented as a generic server framework within vrouter agent that accepts registrations from clients over a Unix domain socket. Anytime a tunnel nexthop is added or deleted, a notification message is sent to each client containing the nexthop and the action (add or delete).

Change-Id: I1c412b062377fe91022d847cf84bd023a4a338bc