initialize deleted_ flag to false in constructor
issue:
------
deleted_ variable was not initialized in unicast mac remote
table constructor, resulting some times in issue where
KSyncObjectManager::Unregister for the object gets triggered
while the object is not yet deleted, when the l2 route table
is deleted it triggers a KSyncObjectManager::Unregister
resulting in a double delete causing segmentatioon fault.

fix:
----
initialize deleted_ flag to false in constructor.
Closes-Bug: 1415928

Change-Id: I381c05975c0fc6ec712effa8466609355238a064