Handle enabling/disabling of DHCP flag and gateway.
If DHCP is enabled for IPAM then dont flood the request coming from
source. This is achieved by not setting DHCP Flood flag in source bridge route.
If DHCP is disabled then flood flag for DHCP is enabled on the bridge
entry of source.

If gateway is not set i.e. 0.0.0.0 is received then dont send in
DHCP reply.
Partial-Bug: #1415014

Change-Id: I86ad76462c212898543ed540f95fb84a010a81a3