Added checks for buffer over-flow
Added check for buffer overflow on server side while encoding
Xmpp OpenConfirm. On detecting overflow the server sends a
TCP close to client and moves Active state to Idle state.

On client side, we check for overflow while encoding Xmpp Open
message. On deteting overflow a TCP close is sent to the server
and the client moves from Connect state to Active State.

Closes GNATS PR 1049616, PR 1048074

Change-Id: Id06344396ccbe53a0c664dfb139fa2ce783559a8
Closes-Bug:1408742