Fix traceback when listing floatingips with LBaaS
Currently when associating a floating-ip to a vip
port we get a traceback since in that case the floating-ip
has back refs to more that one VMI. It always use the first
VMI returned which is not the vip port and could be set
as non user visible.

This patch iterate over all the VMI associate to a fip
a use the one available.

Change-Id: Idf9f913d887aaf1bfef5a24fa94958200eddba33