[Schema] Do not remove VM static route target during initialization
When the schema transformer initializes VN resources, it sets the
routing instance route targets list to the auto-generated route target.
But if the VN already exist and have user defined route targets, it
override it with the auto-generated one, until it received the IF-MAP
notification 'route-target-list'. That could cause a lack of
connectivity.

Closes-Bug: 1418701
(cherry picked from commit 85656cfbe06f40c97c15377e28cb1530124ddbb3)

Conflicts:
	src/config/schema-transformer/to_bgp.py
Change-Id: Ia266961cf6a33a3733f2e8f87534d0b824cd9d8c