Do not remove configured RT from RI on restart
When Schema Transformer restarts, it currently resets all route targets on all
routing instances. If there were any RTs configured on the parent VN, those will
be removed and added back when notification is received for adding those targets
to the VN. This can cause lack of connectivity for some time. Instead, now we
read the configured targets when routing istance is created and add the
configured route targets to the RI.

Change-Id: Id9bdb45c5a782bd4e4ae03a29dd1cd37ca20b081
Closes-Bug: 1418701