Fix Tor-Agent crash for VRF Delete timeout
Issue:
------
when a logical interface is associated with multiple VMIs
and administrator removes one VMI from the list associated
it results in deletion of existing VRF and re-addition
again. By this time if there are any routes imported from
TOR, they hold the reference to VRF and doesnot allow VRF
to clean up, which results in a delete timeout.
In usual scenarios VRF is deleted along with VN, so we
didnot observe this issue earlier.

Fix:
----
Maintain a VRF dependency list in UnicastMacLocalOvsdb
table and trigger re-eval of entries on VRF delete.
which removes the OvsdbEntry for unicast mac local and
re-adds resulting in removing the VRF reference and move
the entry to add-defer state to wait for new VRF object.

This trigger of re-eval is triggered in a work queue to
ensure the order of events such that re-eval kicks in after
the vn_ovsdb_entry becomes in-active to hold the
re-addition of route in Add defer state.

Also adding fix for marking a logical entry as incomplete
if logical interface to vmi association is removed to allow
completion of logical switch delete.

Closes-Bug: 1420903
(cherry picked from commit fed0d74f972857fd1e9d41f1ea33794b81631fab)

Change-Id: Icfe97976f2211d83275edad9da893222ea9d8abf