Check if the VM node is already deleted before adding VR-VM link.
There can be scenarios where the VM node gets deleted i.e. all origin's are
gone but the vm-uuid-mapper has not seen the delete yet. Before he sees the
delete, we receive a VmReg from the agent. We should ignore this VmReg if the
VM is already deleted. Each VM will have a unique/new uuid and receiving a
VmReg from the agent in response to a VM-spawn means that its for the existing
VM and not any new incarnations. For this reason, we should not consider it as
a pending VmReg either.

Closes-Bug: 1421426

Change-Id: I2dc6c570de9a63b6b8597efd1097997e73f02db3