Fix issue with query processing when limit option is specified
Presently, PostProcessingQuery::merge_processing() and
PostProcessingQuery::final_merge_processing() was coded with the
assumption that limit option in the query would be always accompanied
with sort/sort_fields in the query.

In PostProcessingQuery::merge_processing() and
PostProcessingQuery::final_merge_processing(), results from parallel
queries should be merged even if the query is not requested to be sorted.

Change-Id: I096a77efde1797e610acf43e2a1448c71d08684b
Closes-Bug: #1421957