Problem: Walk Done used to get executed in partition where last walk is completed. This used to result in running of walk done in parallel with operations like WalkCancel. Solution: In DB table walker, force walk done to be called from task_instance -1. This ensures that walk done is called in exclusion with same task running in any other instance.
Closes-bug: #1418288
Problem as seen in vnsw agent: Walk done used to reset walk ID stored in
agent. This walk id is used to cancel walk or know status of walk.
Cancel walk in turn calls DBTable Walker WalkCancel. However cancel walk
and walk done because of above mentioned issue can happen in parallel
and as a result CancelWalk can give the walk id which has been reset by walk
done. Result is assert as DBTableWalker is not aware of that walk id.

Change-Id: I45e8f44b8a318280b23b8ee97af6f4b71769bead