Fixing no such device errors for logical intf add
issue:
------
Logical interface and remote physical interface are derived
from interface class which tries to fetch mac addr from
kernel for interface that are actually present on remote
system resulting in failure to find device.

fix:
----
overriding GetOsParams to just initialize the os params to
defaults instead of trying to fetch it from kernel for
logical and remote physical interface.

Change-Id: I20f7c10f0b848196afa50b150c1b46a9b60fb487
Closes-Bug: 1423420