Improve replicator scaling by not walking vpn tables
When the VRF for a new VN is created, the route path replicator needs to
import routes with matching targets from other VRFs and the VPN table.
Existing implementation of replicator achieves this by scheduling a walk
of all VRF tables that have an export target that's the same as one of
the import targets of the new VRF.  Further, it also schedules a walk
of the VPN table to import matching routes received from BGP peers.

Walking entire VPN table for every new VRF can be expensive as the VPN
table could have a very large number of routes, especially if there's a
large number of VRFs.

Address this by implementing the following changes:

- Instead of walking the VPN table when an import targget is added to or
  removed from VRF, trigger notification of dependent VPN routes for the
  target in question.  The list of dependent VPN routes is maintained by
  rtarget group manager already.
- Since we do not walk the VPN table, we can no longer use the walk done
  notification to trigger unregister of the VPN table.  Use a reference
  count on the TableState for the VPN table to decide when it's safe to
  unregister and delete the TableState.
- Simplify code by creating VPN TableState at replicator initialization
  and deleting it only on shutdown.
- VPN TableState can be deleted due to 3 different triggers viz.
  o When the TableState gets a ManagedDelete callback
  o When the last RtGroup leaves the VPN table
  o When the last DBState on VPN table routes is cleared
- Add tests to exercise the 3 triggers mentioned above
- Add unit tests that join and leave import only and export only targets
  for a table.
- Add methods for TableState manipulation to reduce code duplication
- Cleanup to eliminate recursive call to RoutePathReplicator::Join

Change-Id: I5b653bce2dcdc396c2afb58c456496a996e849b4
Closes-Bug: 1425796