Fix Tor-Agent assert for KSyncEntry GetReference
Issue:
------
On connection close delete is triggered for all the KSync
Objects, creation of entry after this deletion in not
allowed.
Here Logical Switch Object is deleted and Unicast Remote
Entry Table is yet to receive a delete, in the meanwhile
a route table entry get notified and tries to take create
a temp entry for reference in Logical Switch Entry which
is already deleted

Fix:
----
Since delete trigger for Unicast Remote Entry Table happens
in context of VRF table walk, which will have a yeild in
between. during this time if a notification for route table
happens, it needs to be ignored.

Also fixing a memory leak in idl deletion path

Change-Id: I692a4b4c277e6b29215e13aad48276d09d3f9377
Closes-Bug: 1424531