Invoke RetryDelete() on unregister of last listener
The lifetime actors for route db-tables will wait till number of entries
in dbtable is 0 and there are no listeners to the table.

DBTable infrastructure invokes RetryDelete() when a DBTablePartition
becomes empty. In addition, it should invoke RetryDelete() when all
db-table listeners are gone also.

The fix will invoke RetryDelete() when number of listeners becomes empty

Change-Id: I659a07bde0b790be08b5cafe1349c3c6b640d1ec
Closes-Bug: #1429348