Optimize Dns config handling in Agent.
Ipam Notify is invoked when link with virtual-network or the link
with virtual-dns changes, which then goes over all the interfaces
to check for impact on DNS dynamic records. With high number of
interfaces and virtual-networks, the processing is done multiple times.

Optimize this by calling only IPAM notify when IPAM properties change.
The link handling is done in virtual-network notify and virtual-dns
handling is done in its own notify.

Change-Id: I2e1d1ad8efb87ab0129dd396b93a647bb83ce9e0