caching only most frequently used connections in vnc_openstack
The vnc_openstack creates and caches the connection object for user
context. If there are more number of users, there will be more number
of connections. The cached information is never freed.

Through this patch, the number of connection to be cache can be
configured. Caching is done only for most frequently used contexts.

Change-Id: I4df370108718419f396923f095511eca146f3f80