Performance optimization for phyiscal-device-vn tables
Old Design:
----------
physical-device-vn table is built from physical-device IFMap node.
We walk thru the following chain to build the table,
physical-device <-> physical-interface <-> logical-interface <-> VMI <-> VN

The dependency manager is configured such that any change to VN or link
in the path above will invoke change on physical-device. The
physical-device will in-turn traverse the link to find new se of
physical-device-vn entries. This leads to scaling issues since we do the
walk multiple times and also invoke IFNodeToReq to physical-device-vn
entry multiple times.

New Design:
----------
The physical-device-vn entry is built from VMI <-> VN link. On seeing
new VMI <-> VN link, we find the physical-device for the VMI and make
the physical-device-vn entry based on this information.

Following tables are maintained to manage the physical-device-vn entries

1. Interface table maintains vmi_to_physical_device_vn_tree_. The tree
   is keyed by VM-Interface UUID and data contains physical-device and
   VN for the interface.

2. Physical-Device-VN table maintains a config-tree. The tree is keyed
   by <dev, vn, vmi>. Multiple VM-Interfaces can contribute to one
   physical-device-vn entry.

   The physical-device-vn entry is created when first entry for a given
   <dev, vn> is added. Subsequent adds to the tree do not generate change
   to physical-device-vn table.

   The physical-device-vn entry is deleted when the last entry with a given
   <dev, vn> value is deleted.

   The physica-device-vn config-tree also tracks change to vxlan-id and
   vn-id. Any change to vxlan-id or vn-id also triggers change to the
   physical-device-vn entry

Change-Id: Ib1c22698983cbe6d8ba6d130d8de122a7c196780