Catch the AddrFormatError exception if ip_addr is None
AddrFormatError exception is seen for "neutron port-list"
in the function _ip_address_to_subnet_id.

The issue is seen because ip_obj.get_instance_ip_address()
is returning None in the function _port_vnc_to_neutron.

This patch doesn't fix the actual issue. It just catches
the AddrFormatError exception so that "neutron port-list"
is successful.

Change-Id: Ica2cc7a30f2b4819f6a33f2be79a2770179e5220
partial-bug: #1432655