Retrieve instance_ip_back_refs when checking subnet deletion
Currently the instance_ip_back_refs of VirtualNetwork is used
to check if a "subnet" has in used ports. Before this
patch the VirtualNetwork object was read without reading this
field instance_ip_back_refs. Without this field retrieved
the subnet is always marked as free of ip instance allocation.

Change-Id: I34ae578719f7e52f3b6c19defdccd7088134b49a