Update the collector column families from Ascii to UTF8
Cassandra returns InvalidRequestException when collector
tries to add a column and if the column data contains
unicode characters.

The issue is seen when any resource is created or deleted and
the resource name has unicode characters.

This patch fixes the issue by changing the collector column
family validation types from AsciiType to UTF8Type.
The patch also takes care of upgrading the existing column families
if they exist in the db.

Change-Id: I54ae9e756f7f67ce684ef9e3ba557038bc8f8731
Closes-bug: #1433021