Handle null value in the FlowRecordTable query response in OverlayToUnderlayMapper utility
In the fix for https://bugs.launchpad.net/juniperopenstack/+bug/1433203,
query-engine has been modified to send null value for column fields which
are not present in the FlowRecordTable.

Problem:
Presently, OverlayToUnderlayMapper utility doesn't validate the value of the flow record fields
in the response and hence sends where clause with null values in the UFlowData query if any of
the column values in the FlowRecord query response contains null. This results in the failure
of query validation in query-engine for UFlowData query.

Fix:
Skip the processing of record/row in the FlowRecord query response if any of the column fields is null.

Change-Id: I7cd5e5352fccff4b482c4090214603c57b1a2d11
Closes-Bug: #1433915