Fix race condition in bgp_xmpp_deferq_test
We used to check that the session on agent is down after processing
a duplicate subscribe or a spurious unsubscribe at the CN.  Some of
these tests fail once in a while because the session goes down and
comes back up before we check the session state.

Fix by checking the flap count on the agent instead of the state.

Change-Id: I434c2a334285e043193fe0f73fd783f69ebe7c18
Closes-Bug: 1431025