Fix incorrect assertion in BgpXmppChannel::MembershipResponseHandler
If the current incarnation of the routing instance is being deleted
and a subsequent subscribe and routes have been sent by the agent for
the new incarnation of the instance, the assertion for the defer q
being empty can fail. The subscribe for the to be re-created instance
is enqueued in vrf_membership_request_map_ whereas the pending routes
are enqueued in the defer queue.

Change-Id: Id11a9d6b0b44f79176a0b83cbf8f89bba476be55
Closes-Bug: 1434383
(cherry picked from commit 95af40ce8f75a807dfdd6d5340821655ab48e91f)