Fix TOR Agent Crash on Bridge Route Notification
Issue:
------
OVSDB client receives Add/change notification for a bridge
route entry in deleted VRF, where it tries to Find a evpn
route in the vrf and results in assertion due to access
of NULL vrf intrustive pointer. as on delete of vrf we
remove vrf reference from all the route tables

Fix:
----
Treat Add/Change notification for a bridge route in deleted
VRF as delete, since we don't reuse vrf object this entry
should anyway be deleted eventually.
Closes-Bug: 1432203

Change-Id: I92d9902108e5ffaa4ece59b792d2f6111804d154