Fix incorrect assertion in BgpXmppChannel::MembershipResponseHandler
If the current incarnation of the routing instance is being deleted
and a subsequent subscribe and routes have been sent by the agent for
the new incarnation of the instance, the assertion for the defer q
being empty can fail. The subscribe for the to be re-created instance
is enqueued in vrf_membership_request_map_ whereas the pending routes
are enqueued in the defer queue.

Change-Id: I68fd12a738f98ec665962a920e7a283dc6bf8706
Closes-Bug: 1434383