Include address as part of key for broadcast MAC routes
Xmpp sessions from multiple TOR-Agents to a CN can use the same IP
address.  Hence, using the source address to construct the RD for
broadcast MAC routes does not guarantee uniqueness of EVPN routes.

Fix is to have the agent set the address field in the broadcast MAC
route to the IP fabric address of the TOR. This makes the broadcast
MAC route from each TOR-Agent unique.

This commit teaches the CN to use the IP address from the broadcast
MAC route.  Agent changes to set the address field to the IP fabric
address of TOR will be committed separately.

Update tests so that mock TOR-Agents behave as described above.

Treat 0.0.0.0/0 the same as 0.0.0.0/32 for backward compatibility
with older versions of agent code.  They set address for broadcast
MAC routes to 0.0.0.0/0. Since we no longer ignore the address for
broadcast MAC routes, we should not treat 0.0.0.0/0 as an invalid
address.

Change-Id: Iafc19d865626e0555cb52c70e9ce6810e9780568
Partial-Bug: 1388932