Add table name of the nodes (of the link) to create the key for links in the link-table.
There can be cases where we add a link between 2 nodes with the same name. This
causes a problem for links with metadata. While adding the link between midnode
and second, we find the link between first and midnode leading to an assert.

Fix is to use the table-name of both the nodes of the link while creating the
link's key.

Change-Id: Ibbaeff7845b74da56abe1488dced24f131d75377
Closes-Bug:1434342