Fix TOR Agent Crash for missing dev info
Issue:
------
During deletion of config, due to difference in the order
of notification events, we have a state where physical
device is deleted but physical device vn entry is still
active, during this we release the refernce of physical
device and fails to process this entry in KSync Notify
resulting in crash.

Fix:
----
Add DB filter to trigger delete for Logical switch entry
if device pointer is missing in DB entry, since we
should anyway receive delete for this entry eventually.

Added test-case for the same.
Closes-Bug: 1438006

Change-Id: I9392da2bf5348850621199e57d5fc1aa5f8bf075