1. Add a “severity” field. This is also are per-UVE, per-alarm-type. This will be an integer between 0 and 7, with 0 as the most severe.
2. Add a “timestamp” field.
This is also per-UVE, per-alarm-type.
It will give microseconds since epoch, and will be updated whenever the alarm description is created/updated.
The client must supply this value when acknowledging alarms.
This is how analytics detects when the client’s view of the alarm is stale and the acknowledgement should not be allowed.

Change-Id: I42c2f3dbd6d3319e9a2cdf0ccc2480472dc827cb
Partial-Bug: 1428271