Handle VLAN property not set in logical-interface
We do not allow modificaiton of vlan-tag in logical-interface. But, in
IFMap messages its possible that VLAN-Tag property is not set. In
such case, we were setting vlan-tag to 0 and then change the value later
when property is received.

Modified agent code to ignore IFMap messages when vlan-tag property is
not set.

Change-Id: I7447a757c7a15367ffb75e881e266379b26c1abc
Fixes-bug: #1437105