Fix vm-interface pointer in logical-interface is NULL
In config logical-interface has links to multiple vm-interface
In oper-db, logical-interface links to one vm-interface.

In logical_interface.cc BuildData() function, the first vm-interface is
picked. However, there is guarantee that the vm-interface picked is
already created before logical-interface is being created.

Modified dependency-tracking to trigger changes to logical-interface
when a vm-interface is created.

Fixes-bug: #1439941
Change-Id: I39ccd40015c69edc1532279efaf4c7aa92f3103e