config-perf: add support for multikey get in useragent kv table
neutron net-list first finds list of network and subnet ids and
then issues subnet list with list of subnet UUIDs. The neutron
backend in subnets_list() was doing as many roundtrips as # of
subnet UUIDs to useragent-kv table to get subnet details.
This is much faster by adding support for RETRIEVE with
multiple keys. Test for the same.

Change-Id: Ie001bcbbb63f4b885e265a1a2815f650a257c1d3
Closes-Bug: #1433979